I Moved player to the package its in in 0.30
