+++ src/com/mojang/minecraft/player/InputHandler.ts
