+++ src/com/mojang/minecraft/player/InputHandlerImpl.ts
